refactor: extract HomeMeetingPreviewSheet from HomeView; dedupe + section RetroactiveSpeakerUpdater
Pure code motion per codebase audit 2026-07-08 wave 2 (spec W2-D).

HomeView.swift (3112 -> 2430 lines):
- Extract the ~680-line HomeMeetingPreviewSheet struct (and its exclusively-owned
  private helper views: HomeMeetingPodcastPlayer, HomePodcastPlayerButton,
  HomeTranscriptFindBar, TranscriptFindHighlight, HomeMeetingTranscriptLineView,
  HomeMeetingSpeakerPill, HomeMeetingSpeakerColor) into a new
  HomeMeetingPreviewSheet.swift file. Widened HomeHitTarget from private to
  internal since it's now used from both files.
- Skipped the optional pagination/grouping helper extraction: the only
  logic-bearing helper (HomeViewModel.groupByDay) is a static func nested
  inside the ObservableObject itself, and pulling it out would mean editing
  the class body -- out of scope for "no view-model rewrite this wave".

RetroactiveSpeakerUpdater.swift (2056 -> 1231 lines), split into:
- +Scanning.swift: file scanning / frontmatter pre-filter / YAML row parsing
- +TranscriptRewrite.swift: full-transcript section rewrite + scoped label
  fallback
- +BreakdownRewrite.swift: speaker breakdown section rewrite + scoped
  fallback + SpeakerBreakdownConsolidator

Deduped applyScopedTranscriptLabelReplacements and
applyScopedBreakdownNameReplacements into one generic applyScopedReplacements
engine (compared line-by-line first -- the two differ in more than
cosmetics: hard-fail vs soft-skip on an unresolvable channel/diarizer target,
when the occurrence-count gate runs, and what a valid replaced-count looks
like -- all now explicit parameters/closures rather than baked into two
copies). Both original names remain as thin wrappers so call sites are
unchanged.

Hoisted the "footerCandidates" markdown section-boundary array, which was
duplicated verbatim in three places (speakerBreakdownContentRange,
rewriteSpeakerBreakdown, and SpeakerBreakdownConsolidator's differently-named
breakdownFooters), into one private breakdownSectionFooterCandidates
constant in +BreakdownRewrite.swift.

Access widened from private to internal only where a moved function is now
called cross-file (documented per-symbol in the corresponding diff hunks);
everything with a single in-file caller stayed private.

Neither original file appears in scripts/entrypoints/run-tests.sh's
APP_SOURCES list or Tests/FastTests.manifest (the latter lists test files,
not source files), so no manifest updates were needed for the new files.

Local build/tests: build-deps.sh --force was queued behind other worktrees'
builds on the shared flock for the full session and did not finish in time --
CI is the gate for this PR. Static verification done instead: balanced-brace
check on every touched/new file, and a duplicate-declaration sweep across
all four RetroactiveSpeakerUpdater* files turned up only expected local-name
collisions and legitimate overloads.
